fix Republic of Congo being annexed upon winning the civil war
Think these were just the wrong way around.  Previously, even after winning the war, the Republic of Congo would immediately be annexed by the Free Republic as soon as the pop-up appears stating that the Free Republic has been defeated by and capitulated to the Republic. Swapping it around seems to fix it for me.