Six Day War + UAR focus tree fixes
This is a very complex and consequently very buggy event chain, so I've annotated it to make future troubleshooting & editing easier and I've also made a few fixes. The phrasing of some of the annotations is slightly inconsistent, but that's phrasing copied and pasted from the comments in the localisation file.

The main changes are:
- the war will start a week after Egypt closes the Straits of Tiran via its focus tree - as it did historically - instead of always occurring on the same date.
- Egypt/UAR's focuses are now unlocked by the war (The Six-Day War when the war begins, Partial Liberation of Palestine when the UAR wins, an Naksah when it ends either in stalemate or Israeli victory) - previously there seemed to be no way of unlocking them at all
- the Golan Heights and West Bank sections of the conflict should now still occur if Egypt/UAR has annexed the territory of Syria and/or Jordan
- switched Syria's victory national focus to be attached with event .35 and edited event .5 so that it doesn't fire for Syria - the event was kind of redundant for Syria anyway, and having the two border conflicts share the event caused issues if one country owned both territories
- SIX_DAYS_WAR.3 is fixed to use the correctprovinces, so that border conflict can actually occur
-  SIX_DAYS_WAR.14 uses the description from SIX_DAYS_WAR.8.d, but it says you are being given a choice, even though SIX_DAYS_WAR.14 doesn't offer a choice, so the text of desc = SIX_DAYS_WAR.3.d fits better.

Known unresolved issues:
- SIX_DAYS_WAR.192 - Israeli Perspective Counter-Attack into Sinai - appears to have no trigger
- the news events are still a bit of a mess and i've not tried to fix them - some countries don't get properly notified when the war has ended (the original event chain is primarily written for Israel) or get notifications that are clearly written from the perspective of a different country
- if Jordan owns Syria's territory, or vice versa, it won't fire both border conflicts
- if an Arab power other than UAE (e.g. Iraq as the Arab Federation or a unified Arab League) owns Jordan's or Syria's territory, their border conflicts won't fire
- if Israel declines to attack, Egypt/UAE attacks alone only on the Sinai front - when historically the other Arab powers likely would have joined in
- UAR's focus description for UAR_an_Naksah describes the war as a crushing defeat, but the focus still applies even if the war is a stalemate
- UAR's focus UAR_The_Six_Days_War describes the war as an Israeli surprise attack, but it's possible for the war to play out as an Egyptian/UAE surprise attack instead.

And tbh there's probably still a lot of unresolved bugs I've missed still. I've tested it a lot of times, with different combinations of countries involved, and different paths, but it's a very complex chain and there are a lot of different possible combinations, so I've not tested exhaustively. And I'm not an experienced modder, so I'm sure things have slipped through the cracks, but it seems significantly less buggy now than before. There is probably an argument for rewriting the entire war, getting rid of all the complex chains of Border Conflicts, and just having it as a regular war with scripted peace deals, but that's above my pay grade - I was just trying to make what already exists functional.