switched heavy naval guns to do gun damage
As far as I'm aware, the only ships in this period to use heavy guns were the American Iowa class battleships, and my understanding is that they were only kept in service to be used for shorebombardment and never used for naval combat. In the mod we have LIGHT ATTACK set up to be effective at shore bombardment, and HEAVY ATTACK set up to represent long-range anti-ship missiles. It doesn't really make sense for an Iowa class battleship to be weak at shore bombardment or to be able to match the range of a missile cruiser, so I've switched the heavy guns to use LIGHT ATTACK. The downside of this is that they will probably be unrealistically effective at fighting light ships - but I'm not sure it's possible to resolve this in a way that will leave them with realistic effectiveness against light ships, heavy ships, and shore bombardment.